+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+
+- opensearch: Add Opensearch as new product with version `3.1.0` ([#1215]).
+
+[#1215]: https://github.com/stackabletech/docker-images/pull/1215
+
 ## [25.7.0] - 2025-07-23
 
 ## [25.7.0-rc1] - 2025-07-18
@@ -65,7 +71,6 @@ All notable changes to this project will be documented in this file.
 - hadoop: backport HADOOP-19352, HADOOP-19335, HADOOP-19465, HADOOP-19456 and HADOOP-19225 to fix vulnerabilities in Hadoop `3.4.1` ([#1184])
 - hadoop: Backport HADOOP-18583 to make OpenSSL 3.x work with the native hadoop libraries ([#1209]).
 - spark: backport [SPARK-51311] Promote bcprov-jdk18on to compile scope ([#1212]).
-- opensearch: Add Opensearch as new product with version `3.1.0` ([#1215]).
 
 ### Changed
 
@@ -230,7 +235,6 @@ All notable changes to this project will be documented in this file.
 [#1197]: https://github.com/stackabletech/docker-images/pull/1197
 [#1209]: https://github.com/stackabletech/docker-images/pull/1209
 [#1212]: https://github.com/stackabletech/docker-images/pull/1212
-[#1215]: https://github.com/stackabletech/docker-images/pull/1215
 
 ## [25.3.0] - 2025-03-21